Refactor classes in storage operations
